Physics update and timing changes, std::string to string_t changes.

+++ library/src/breathe/vehicle/cVehicle.cpp
@@ -13,6 +13,7 @@
 
 #include <breathe/breathe.h>
 
+#include <breathe/util/cString.h>
 
 #include <breathe/math/math.h>
 #include <breathe/math/cVec2.h>
@@ -34,9 +35,6 @@
 #include <breathe/game/cLevel.h>
 
 #include <breathe/physics/physics.h>
-#include <breathe/physics/cContact.h>
-#include <breathe/physics/cRayCast.h>
-#include <breathe/physics/cPhysicsObject.h>
 
 #include <breathe/game/cPlayer.h>
 #include <breathe/game/cPetrolBowser.h>
@@ -604,7 +602,7 @@ namespace breathe
 
 		void cVehicle::FillUp(cPetrolBowser *pBowser)
 		{
-			float fSpace = static_cast<unsigned int>(fPetrolTankSize);
+			float fSpace = fPetrolTankSize;
 			for(int i=0;i<PETROL_SIZE;i++)
 				fSpace-=static_cast<unsigned int>(vPetrolTank[i]);
